backend/requirements.txt
absl-py==2.1.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.11.1
asttokens==3.0.1
astunparse==1.6.3
asyncpg==0.31.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.3
databases==0.9.0
debugpy==1.8.19
decorator==5.2.1
distro==1.9.0
dj-database-url==3.1.1
Django==6.0.2
django-cors-headers==3.11.0
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
executing==2.2.1
fastapi==0.128.0
Flask==3.1.0
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
greenlet==3.3.0
grpcio==1.68.1
gunicorn==25.1.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.0
idna==3.10
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.8.0
jupyter_core==5.9.1
keras==3.7.0
langchain-core==1.2.7
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.2.1
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
nest-asyncio==1.6.0
numpy==2.0.2
openai==1.56.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
orjson==3.11.5
ormsgpack==1.12.1
packaging==24.2
parso==0.8.5
platformdirs==4.5.1
prompt_toolkit==3.0.52
protobuf==5.29.1
psutil==7.2.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.3.0
python-dateutil==2.9.0.post0
pytz==2021.3
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.50.0
tenacity==9.1.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2021.5
urllib3==2.2.3
uuid_utils==0.13.0
uvicorn==0.40.0
wcwidth==0.3.2
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.11.0
wrapt==1.17.0
xxhash==3.6.0
zstandard==0.25.0
